requirements.txt
tqdm
future
scipy
numpy
sklearn
pillow
requests
six
tarfile
zipfile
